requirements.txt
streamlit
pandas
matplotlib
requests
python-dotenv
PyMuPDF
pdfplumber
seaborn

